python_container/requirements.txt
Flask
beautifulsoup4
requests
python-dotenv